+++ packages/contracts/contracts/voting/majority/MajorityVotingBase.sol
@@ -151,12 +151,10 @@ abstract contract MajorityVotingBase is
     /// @param abstain The number of abstain votes casted.
     /// @param yes The number of yes votes casted.
     /// @param no The number of no votes casted.
-    /// @param totalVotingPower The total voting power available at the block prior to the proposal creation.
     struct Tally {
         uint256 abstain;
         uint256 yes;
         uint256 no;
-        uint256 totalVotingPower;
     }
 
     /// @notice The [ERC-165](https://eips.ethereum.org/EIPS/eip-165) interface ID of the contract.
@@ -310,7 +308,9 @@ abstract contract MajorityVotingBase is
         return
             (RATIO_BASE - proposal_.parameters.supportThreshold) * proposal_.tally.yes >
             proposal_.parameters.supportThreshold *
-                (proposal_.tally.totalVotingPower - proposal_.tally.yes - proposal_.tally.abstain);
+                (totalVotingPower(proposal_.parameters.snapshotBlock) -
+                    proposal_.tally.yes -
+                    proposal_.tally.abstain);
     }
 
     /// @inheritdoc IMajorityVoting
@@ -349,6 +349,11 @@ abstract contract MajorityVotingBase is
     function votingMode() public view virtual returns (VotingMode) {
         return votingSettings.votingMode;
     }
+
+    /// @notice Returns the total voting power checkpointed for a specific block number.
+    /// @param _blockNumber The block number.
+    /// @return The total voting power.
+    function totalVotingPower(uint256 _blockNumber) public view virtual returns (uint256);
 
     /// @notice Returns all information for a proposal vote by its ID.
     /// @param _proposalId The ID of the proposal.