+++ packages/subgraph/tests/addresslist-voting/addresslist-voting.test.ts
@@ -33,7 +33,11 @@ import {
   SNAPSHOT_BLOCK,
   TOTAL_VOTING_POWER
 } from '../constants';
-import {createDummyActions, createGetProposalCall} from '../utils';
+import {
+  createDummyActions,
+  createGetProposalCall,
+  createTotalVotingPowerCall
+} from '../utils';
 import {
   createNewAddressesAddedEvent,
   createNewVoteCastEvent,
@@ -75,9 +79,14 @@ test('Run AddresslistVoting (handleProposalCreated) mappings with mock event', (
     '0', // abstain
     '0', // yes
     '0', // no
-    TOTAL_VOTING_POWER,
 
     actions
+  );
+
+  createTotalVotingPowerCall(
+    CONTRACT_ADDRESS,
+    SNAPSHOT_BLOCK,
+    TOTAL_VOTING_POWER
   );
 
   // create event
@@ -223,9 +232,14 @@ test('Run AddresslistVoting (handleVoteCast) mappings with mock event', () => {
     '0', // abstain
     '1', // yes
     '0', // no
-    TOTAL_VOTING_POWER,
 
     actions
+  );
+
+  createTotalVotingPowerCall(
+    CONTRACT_ADDRESS,
+    SNAPSHOT_BLOCK,
+    TOTAL_VOTING_POWER
   );
 
   // create event
@@ -281,7 +295,6 @@ test('Run AddresslistVoting (handleVoteCast) mappings with mock event', () => {
     '0', // abstain
     '2', // yes
     '0', // no
-    TOTAL_VOTING_POWER,
 
     actions
   );
@@ -342,7 +355,6 @@ test('Run AddresslistVoting (handleVoteCast) mappings with mock event and vote o
     '0', // abstain
     '0', // yes
     '0', // no
-    TOTAL_VOTING_POWER,
 
     actions
   );